moved Recs to pages directory, styling Recs component
